+++ src/App.css
@@ -1,7 +1,23 @@
-.App{
+.Intro{
   min-height: 100vh;
   background-image: url("./childe.jpg");
-  background-color: rgb(234, 192, 192)
+  background-color: rgb(234, 192, 192);
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.Home{
+  min-height: 100vh;
+  width:100%;
+  background-image: url("./childe.jpg");
+  background-color: rgb(234, 192, 192);
+}
+
+.Intro-header{
+  padding: 30px;
 }
 
 .search-form{
